Implement P-384 and P-521
Clean up the SPAKE plugin and implement new curves.

I have tried to split up the code to improve auditability.

At the core is iana.[ch] which contains the contents of the IANA registry.
On top of this we have groups.[ch] which implements the interface for group
implementations. It references openssl.[ch] which implements the actual
groups.

In addition to this, we have util.[ch] and trace.h which implement
shared utility functions and tracing macros (respectively).

The client and server implementations of SPAKE refer to just groups.h,
util.h and trace.h.